# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Simple CRUD Backend API
  version: 1.0.0
  description: |
    Training API for students. Includes register, login, and authenticated profile endpoint.
servers:
  - url: http://localhost:4000
    description: Local
  - url: https://your-render-service.onrender.com
    description: Render
tags:
  - name: Auth
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation failed
        '409':
          description: Email already exists
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation failed
        '401':
          description: Invalid credentials
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/PublicUser'
        '401':
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
          example: Student One
        email:
          type: string
          format: email
          example: student@example.com
        password:
          type: string
          minLength: 8
          example: mySecurePassword
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: student@example.com
        password:
          type: string
          example: mySecurePassword
    PublicUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful
        token:
          type: string
        user:
          $ref: '#/components/schemas/PublicUser'